THRIFT-1325 Thrift SVN and latest GCC issue: Undefined symbols: "apache::thrift::protocol::TBinaryProtocolT<apache::thrift::transport::TTransport>::VERSION_MASK" Patch: Brandon Arp
git-svn-id: https://svn.apache.org/repos/asf/thrift/trunk@1211842 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/cpp/src/protocol/TBinaryProtocol.h
@@ -36,8 +36,8 @@ template <class Transport_>
 class TBinaryProtocolT
   : public TVirtualProtocol< TBinaryProtocolT<Transport_> > {
  protected:
-  static const int32_t VERSION_MASK = 0xffff0000;
-  static const int32_t VERSION_1 = 0x80010000;
+  static const int32_t VERSION_MASK = ((int32_t)0xffff0000);
+  static const int32_t VERSION_1 = ((int32_t)0x80010000);
   // VERSION_2 (0x80020000)  is taken by TDenseProtocol.
 
  public:
